+++ src/Metadata/PropertyTypeUnknown.php
@@ -4,16 +4,18 @@ declare(strict_types=1);
 
 namespace Liip\MetadataParser\Metadata;
 
+use Symfony\Component\TypeInfo\Type;
+use Symfony\Component\TypeInfo\Type\BuiltinType;
+use Symfony\Component\TypeInfo\TypeIdentifier;
+
+/**
+ * @extends AbstractPropertyType<BuiltinType<TypeIdentifier::MIXED>, bool>
+ */
 final class PropertyTypeUnknown extends AbstractPropertyType
 {
     public function __construct(bool $nullable)
     {
-        parent::__construct($nullable);
-    }
-
-    public function __toString(): string
-    {
-        return 'mixed';
+        parent::__construct(Type::mixed(), $nullable);
     }
 
     public function merge(PropertyType $other): PropertyType
@@ -22,6 +24,6 @@ final class PropertyTypeUnknown extends AbstractPropertyType
             throw new \UnexpectedValueException(\sprintf('Can\'t merge type %s with %s, they must be the same', self::class, $other::class));
         }
 
-        return new self($this->isNullable() && $other->isNullable());
+        return new self($this->nullable && $other->isNullable());
     }
 }