+++ tests/Metadata/PropertyTypeIterableTest.php
@@ -6,18 +6,21 @@ namespace Tests\Liip\MetadataParser\Metadata;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
+use Liip\MetadataParser\Metadata\PropertyType;
 use Liip\MetadataParser\Metadata\PropertyTypeClass;
 use Liip\MetadataParser\Metadata\PropertyTypeIterable;
 use Liip\MetadataParser\Metadata\PropertyTypePrimitive;
 use PHPUnit\Framework\Attributes\Small;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\TypeInfo\Type;
 
 #[Small]
 class PropertyTypeIterableTest extends TestCase
 {
     public function testNestedArrayLeaf(): void
     {
-        $propertyType = new PropertyTypeIterable(new PropertyTypeIterable(new PropertyTypePrimitive('int', false), false, false), false, false);
+        $inner = self::list(new PropertyTypePrimitive(Type::builtin('int'), false));
+        $propertyType = self::list($inner);
 
         $this->assertInstanceOf(PropertyTypeIterable::class, $propertyType->getSubType());
         $this->assertInstanceOf(PropertyTypePrimitive::class, $propertyType->getSubType()->getSubType());
@@ -26,16 +29,16 @@ class PropertyTypeIterableTest extends TestCase
 
     public function testDefaultListIsNotCollection(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $list = new PropertyTypeIterable($subType, false, false);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $list = self::list($subType);
 
         $this->assertFalse($list->isTraversable());
     }
 
     public function testConcreteCollectionClassIsKept(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $collection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $collection = self::list($subType, ArrayCollection::class);
 
         $this->assertTrue($collection->isTraversable());
         $this->assertSame(ArrayCollection::class, $collection->getTraversableClass());
@@ -43,18 +46,18 @@ class PropertyTypeIterableTest extends TestCase
 
     public function testMergeListWithClassCollection(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $list = new PropertyTypeIterable($subType, false, false);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $list = self::list($subType);
 
         $this->expectException(\UnexpectedValueException::class);
-        $list->merge(new PropertyTypeClass(Collection::class, true));
+        $list->merge(new PropertyTypeClass(Type::object(Collection::class), true));
     }
 
     public function testMergeInterfaceCollectionListWithClassCollection(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $typeHintedCollection = new PropertyTypeClass(Collection::class, true);
-        $defaultCollection = new PropertyTypeIterable($subType, false, false, Collection::class);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $typeHintedCollection = new PropertyTypeClass(Type::object(Collection::class), true);
+        $defaultCollection = self::list($subType, Collection::class);
 
         $result = $defaultCollection->merge($typeHintedCollection);
         $this->assertInstanceOf(PropertyTypeIterable::class, $result);
@@ -66,9 +69,9 @@ class PropertyTypeIterableTest extends TestCase
 
     public function testMergeClassInterfaceCollectionWithConcreteCollectionList(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $typeHintedCollection = new PropertyTypeClass(Collection::class, true);
-        $explicitCollection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $typeHintedCollection = new PropertyTypeClass(Type::object(Collection::class), true);
+        $explicitCollection = self::list($subType, ArrayCollection::class);
 
         $result = $typeHintedCollection->merge($explicitCollection);
         $this->assertInstanceOf(PropertyTypeIterable::class, $result);
@@ -80,9 +83,9 @@ class PropertyTypeIterableTest extends TestCase
 
     public function testMergeConcreteCollectionListWithClassInterfaceCollection(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $typeHintedCollection = new PropertyTypeClass(Collection::class, true);
-        $explicitCollection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $typeHintedCollection = new PropertyTypeClass(Type::object(Collection::class), true);
+        $explicitCollection = self::list($subType, ArrayCollection::class);
 
         $result = $explicitCollection->merge($typeHintedCollection);
         $this->assertInstanceOf(PropertyTypeIterable::class, $result);
@@ -94,9 +97,9 @@ class PropertyTypeIterableTest extends TestCase
 
     public function testMergeConcreteCollectionListWithInterfaceCollectionList(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $defaultCollection = new PropertyTypeIterable($subType, false, false, Collection::class);
-        $explicitCollection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $defaultCollection = self::list($subType, Collection::class);
+        $explicitCollection = self::list($subType, ArrayCollection::class);
 
         $result = $explicitCollection->merge($defaultCollection);
         $this->assertInstanceOf(PropertyTypeIterable::class, $result);
@@ -108,9 +111,9 @@ class PropertyTypeIterableTest extends TestCase
 
     public function testMergeDefaultCollectionListWithConcreteCollectionList(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $defaultCollection = new PropertyTypeIterable($subType, false, false, Collection::class);
-        $explicitCollection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $defaultCollection = self::list($subType, Collection::class);
+        $explicitCollection = self::list($subType, ArrayCollection::class);
 
         $result = $defaultCollection->merge($explicitCollection);
         $this->assertInstanceOf(PropertyTypeIterable::class, $result);
@@ -122,9 +125,9 @@ class PropertyTypeIterableTest extends TestCase
 
     public function testMergeListAndInterfaceCollectionList(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $list = new PropertyTypeIterable($subType, false, false);
-        $collection = new PropertyTypeIterable($subType, false, false, Collection::class);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $list = self::list($subType);
+        $collection = self::list($subType, Collection::class);
 
         foreach ([$list->merge($collection), $collection->merge($list)] as $result) {
             $this->assertInstanceOf(PropertyTypeIterable::class, $result);
@@ -137,9 +140,9 @@ class PropertyTypeIterableTest extends TestCase
 
     public function testMergeListWithConcreteCollectionList(): void
     {
-        $subType = new PropertyTypePrimitive('int', false);
-        $list = new PropertyTypeIterable($subType, false, false);
-        $collection = new PropertyTypeIterable($subType, false, false, ArrayCollection::class);
+        $subType = new PropertyTypePrimitive(Type::builtin('int'), false);
+        $list = self::list($subType);
+        $collection = self::list($subType, ArrayCollection::class);
 
         foreach ([$list->merge($collection), $collection->merge($list)] as $result) {
             $this->assertInstanceOf(PropertyTypeIterable::class, $result);
@@ -148,5 +151,15 @@ class PropertyTypeIterableTest extends TestCase
             $this->assertSame(ArrayCollection::class, $result->getTraversableClass());
             $this->assertSame((string) $list->getSubType(), (string) $result->getSubType());
         }
+    }
+
+    private static function list(PropertyType $subType, ?string $traversableClass = null): PropertyTypeIterable
+    {
+        return new PropertyTypeIterable(
+            Type::list($subType->getTypeInfo()),
+            false,
+            $subType,
+            $traversableClass,
+        );
     }
 }