+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Remove hard dependency on `doctrine/annotations`
   Attributes are out for a while now, so we drop the hard dependency on the annotation library. In case it is still
   required, it can still be installed and used by passing the `Reader` to the `JMSParser`.
+* Directly create types and pass them to the `PropertyType*` classes instead of creating them inside the class
 
 # Version 2.x